Split the version metric out to its own package
