+++ hack/.linted_packages
@@ -160,6 +160,7 @@ pkg/util/logs
 pkg/util/maps
 pkg/util/validation/field
 pkg/util/workqueue
+pkg/version/prometheus
 pkg/volume
 pkg/volume/downwardapi
 pkg/volume/quobyte
